+++ packages/migration/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.1.6 - 2020-08-13
+
+- added `--dry-run` option to run pending migrations without commiting them to quickly check for errors
+
+- Dependencies bump:
+  - `@fewlines/fwl-database`: `1.1.0` => `@fwl/database`: `0.1.1` (Package was renamed)
+
 ## 0.1.5 - 2020-07-27
 
 - Dependencies bump: